+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # UNRELEASED
 
+* IntuneAppCategory
+  * Initial release.
 * AADApplication
   * Added AppRoles
   * Added AuthenticationBehavior
@@ -49,6 +51,9 @@
 * EXOMailboxFolderPermission
   * Initial Release.
 * EXOMailboxIRMAccess
+  * Initial Release.
+* EXOMailTips
+  * Remove property `Ensure` since this resource is of type `IsSingleInstance`
 * EXOManagementScope
   * Initial Release.
 * EXORetenionPolicy
@@ -93,6 +98,8 @@
   * Added support for Auto Labeling settings
     FIXES [#3784](https://github.com/microsoft/Microsoft365DSC/issues/3784)
 * SentinelSetting
+  * Initial release.
+* SentinelWatchlist
   * Initial release.
 * SPOAccessControlSettings
   * Added support for property EnableRestrictedAccessControl.